requirements.txt
absl-py==2.3.1
aiofiles==24.1.0
annotated-types==0.7.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
attrs==25.4.0
cffi==2.0.0
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.2
cycler==0.12.1
Cython==3.1.4
flatbuffers==25.9.23
fonttools==4.60.1
iniconfig==2.1.0
jax==0.7.2
jaxlib==0.7.2
kiwisolver==1.4.9
markdown-it-py==4.0.0
matplotlib==3.10.6
mdurl==0.1.2
mediapipe==0.10.14
ml_dtypes==0.5.3
mypy_extensions==1.1.0
numpy==2.3.3
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
oracledb==3.3.0
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.4.0
pluggy==1.6.0
protobuf==4.25.8
pycparser==2.23
pydantic==2.11.10
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.5
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
rich==14.1.0
scipy==1.16.2
sentencepiece==0.2.1
setuptools==80.9.0
six==1.17.0
sounddevice==0.5.2
typing-inspection==0.4.2
typing_extensions==4.15.0
uv==0.8.23
